% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11test-package.R
\docType{package}
\name{cpp11test-package}
\alias{cpp11test}
\alias{cpp11test-package}
\title{cpp11test: A test suite and benchmark code for 'cpp11'}
\description{
Provides a test suite and benchmarking code for the 'cpp11' package.
}
\author{
\strong{Maintainer}: Jim Hester \email{james.f.hester@gmail.com} (\href{https://orcid.org/0000-0002-2739-7082}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
