% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11.R
\name{roxcpp2_}
\alias{roxcpp2_}
\title{Roxygenise C++ function II}
\usage{
roxcpp2_(x)
}
\arguments{
\item{x}{numeric value}
}
\description{
Dummy function to test roxygen2. It adds 2.0 to a double.
}
\examples{
roxcpp2_(1.0)
}
