% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11.R
\name{roxcpp3_}
\alias{roxcpp3_}
\title{Roxygenise C++ function III}
\usage{
roxcpp3_(x)
}
\arguments{
\item{x}{numeric value}
}
\description{
Dummy function to test roxygen2. It adds 3.0 to a double.
}
\examples{
roxcpp3_(1.0)
}
