% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11.R
\name{roxcpp4_}
\alias{roxcpp4_}
\title{Roxygenise C++ function IV}
\usage{
roxcpp4_(x)
}
\arguments{
\item{x}{numeric value}
}
\description{
Dummy function to test roxygen2. It adds 4.0 to a double.
}
\examples{
roxcpp4_(1.0)
}
