% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11.R
\name{roxcpp5_}
\alias{roxcpp5_}
\title{Roxygenise C++ function V}
\usage{
roxcpp5_(x)
}
\arguments{
\item{x}{numeric value}
}
\description{
Dummy function to test roxygen2. It adds 5.0 to a double.
}
\examples{
roxcpp5_(1.0)
}
