% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11.R
\name{roxcpp7_}
\alias{roxcpp7_}
\title{Roxygenise C++ function VII}
\usage{
roxcpp7_(x)
}
\arguments{
\item{x}{numeric value}
}
\description{
Dummy function to test roxygen2. It adds 7.0 to a double.
}
\examples{
roxcpp7_(1.0)
}
