% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{run_tests}
\alias{run_tests}
\title{Run the cpp11 tests}
\usage{
run_tests(reporter = testthat::default_reporter())
}
\arguments{
\item{reporter}{Reporter to use to summarise output. Can be supplied
as a string (e.g. "summary") or as an R6 object
(e.g. \code{SummaryReporter$new()}).

See \link[testthat]{Reporter} for more details and a list of built-in reporters.}
}
\description{
Run the cpp11 tests
}
